+++ checkpoint/CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - #v1 Add LeafHandlerRegistry and make it as default for V1 codes
 - Modify `loading.load_pytree()` method to automatically detect and handle
 different formats
+- #v1 Adjust v0/v1 compatibility with CheckpointLayout loading logic.
+Require users to explicitly configure a layout if it is not Orbax.
+
 ## [0.11.19] - 2025-07-08
 
 ### Added